fix(lightbake): moucheté per-texel du direct + peinture satinée des murs
Le mur uni restant montrait un grain sableux : bruit per-texel du DIRECT
NEE, exposé par le split (l'ancien flou global 3 passes le masquait,
l'adoucissement à 1 texel non). Murs = couleur unie (pas de map) donc
100 % lightmap.

- neeSamples 16 → 32 : bruit direct ÷√2 à la source.
- Adoucissement du direct au combine : 3×3 → 5×5 (≈2 cm à 2048) — les
  pénombres restent nettes, le moucheté disparaît.
- SHARED_WALL_MATERIAL : #d4b080 (tan orangé) → #cfc2a8 (greige chaud),
  roughness 0.38 → 0.42 — « mur lisse peinture satinée » (demande user) ;
  le tan saturé lisait « sale » sous la GI chaude.

A/B validé sur la pose-mur du feedback : surface satinée propre,
dégradé continu, halos nets.

+++ src/lib/lightbake/radiosityBake.ts
@@ -326,8 +326,9 @@ export async function radiosityBake(
 
     // 3. COMBINE → rtFinal (target DÉDIÉE, stable à travers les re-bakes : les bindings runtime/probe
     //    référencent toujours cette texture, peu importe la parité des passes intermédiaires).
-    //    Le direct est adouci d'1 texel à la volée (anti-aliasing du NEE), il reste net.
-    const cDir = blurFn({ fragUv: uv(), res: float(opts.resolution), r: float(1), src: texture(rtDirect.texture) })
+    //    Le direct est adouci en 5×5 (≈2 cm à 2048) à la volée — tue le moucheté per-texel du NEE
+    //    (visible sur les murs unis, feedback 10/06) en gardant les pénombres nettes.
+    const cDir = blurFn({ fragUv: uv(), res: float(opts.resolution), r: float(2), src: texture(rtDirect.texture) })
     const cInd = texture(wSrc.texture)
     const combineMat = new THREE.MeshBasicNodeMaterial()
     combineMat.colorNode = cDir.add(cInd)
